* ParaViewITKPlugins\CMakeLists: Added new filters
* MedianImageFilter: Changed radius parameter type from int to double

+++ MedianImageFilter/vtkMedianImageFilter.h
@@ -56,14 +56,18 @@ public:
   //void SetStandardDeviation(double sigma0, double sigma1, double sigma2);
   //void SetStandardDeviation(double sigma[3]);
 
-  vtkSetVector3Macro(NeighborhoodRadius, int);
-  vtkGetVector3Macro(NeighborhoodRadius, int);
+//  vtkSetVector3Macro(NeighborhoodRadius, int);
+//  vtkGetVector3Macro(NeighborhoodRadius, int);
+
+  vtkSetVector3Macro(NeighborhoodRadius, double);
+  vtkGetVector3Macro(NeighborhoodRadius, double);
 
 protected:
   vtkMedianImageFilter();
   ~vtkMedianImageFilter();
 
-  int NeighborhoodRadius[3];
+//  int NeighborhoodRadius[3];
+  double NeighborhoodRadius[3];
 
   vtkImageExport*                VTKExporter;
   //BTX
